requirements.txt
fastmcp
requests
fastapi
uvicorn
firecrawl-py